snapshot-controller: Skip unmanaged VolumeSnapshots in cache
This patch adds the functionality to skip VolumeSnapshots that
are not present on the cluster by may be found in the cache awaiting
eviction.

Earlier the check for ODF managed resource was only done when the object
was found in the cluster. We should not be processing any
VolumeSnapshots as they are to be handled by upstream snapshotter.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>