common-controller: Fix nil pointer deref
The snapshot contents that are not supposed to be reconciled
by our snapshotter were incorrectly added to the work queue.

Which led to runtime panic when deleting the snapshot content.

This patch fixes the above issue by adding snapshot content to queue
only when the resource is to be managed by us.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>