common-controller: fix `VolumeGroupSnapshotContent is missing` error
The function `syncUnreadyGroupSnapshot` calls `createGroupSnapshotContent` if a
`VolumeGroupSnapshotContent` is not found in the cache.

The called function has a bug where it returns a stale object with some
required metadata missing in cases where a `VolumeGroupSnapshotContent`
is already present. This object is not cached into the local store as
the `ResourceVersion` field is unset, failing the update object call.

As a result of that, subsequent `syncReadyGroupSnapshot` calls fail
to find the said `VolumeGroupSnapshotContent` as it was not updated in
the cache properly.

This patch fixes it by fetching the `VolumeGroupSnapshotContent` from the `Lister`
instead of `Cache` to always get the most up-to-date version.

Note: This approach ensures we have minimal changes in our fork.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>
(cherry picked from commit 8e7e7991e8b870c4d3223a09e98cefff4d42d533)