common-controller: use lister from the informer to fetch VGSC
This patch ditches the generated clientset and uses the lister
provided by `VolumeGroupSnapshotContent` informer. That way, we
don't query API server and have most up-to-date version of the object
at any point of time.

Doing this also allows us to use standard k8s error checking. The generated
clientset returns an empty object when nothing is found, which is really
weird and leads to all sort of problems, `IsNotFound` check fails.

This patch would also need modifications to inject the VGSC on the fake
controller via a custom lister.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>