common-controller: fix stale `VolumeGroupSnapshotHandle` in `VolumeSnapshotContent`
This patch fixes an issue where the VolumeSnapshot would end up
in a reconcile loop due to the controller not being able to find
the VolumeGroupSnapshot that was set in the VolumeSnapshotContent's
status for a VolumeSnapshot.

The issue arose from a logic error which led to the
VolumeSnapshotContent.Status.VolumeGroupSnapshotHandle never
being updated except for the first/initial creation.

More specifically, the resource being passed to patch was empty
in the cases where VolumeSnapshotContent already existed on the cluster.

This patch passes the correct name to patch function to fix the issue.

Signed-off-by: Niraj Yadav <niryadav@redhat.com>